#120: add addParentNode, addChildNode to pointlinejs for call programmaticaly. Fix add child node for stackChildren parent node
